requirements.txt
wheel==0.43.0
# === hugging face ===
transformers==4.41.2
accelerate==0.31.0
datasets==2.20.0
trl==0.9.4
bitsandbytes
peft==0.11.1
deepspeed==0.14.3
# === useful ===
hydra-core==1.3.2
wandb==0.17.5
pillow==10.3.0
scikit-learn==1.5.0
tensorboard==2.17.0
rich==13.7.1
ruff==0.5.4
torchsummary
# from deepseek-vl
timm
sentencepiece
attrdict
einops==0.8.0
# flash-attn
jupyterlab
# uv pip install torch==2.3.1+cu118 torchvision==0.18.1+cu118 --index-url https://download.pytorch.org/whl/cu118 
# uv pip install torch==2.3.0+cu118 torchvision==0.18.0+cu118 --index-url https://download.pytorch.org/whl/cu118 

# from minerl
# uv pip install env/minerl --no-cache-dir --no-build-isolation
# av
# gym>=0.19.0,<=0.23.1
# opencv-python>=4.1.0.25
# typing>=3.6.6
# lxml>=4.3.3
# Pyro4>=4.76
# getch>=1.0; sys_platform != 'win32' and sys_platform != 'cygwin'
# coloredlogs>=10.0
# daemoniker>=0.2.3
# xmltodict==0.12.0
# inflection>=0.3.1
# imagehash>=4.0.0
# flaky
# pyglet
